#25 Export logic implementation with Export as PDF button
