+++ src/app/components/calculate/calculate.component.ts
@@ -1,5 +1,7 @@
 import { Component, HostListener, Inject, OnInit } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
+import * as jsPDF from 'jspdf';
+import html2canvas from 'html2canvas';
 
 @Component({
   selector: 'app-calculate',
@@ -61,5 +63,23 @@ export class CalculateComponent implements OnInit {
 
   refreshPage() {
     this.doc.defaultView?.location.reload();
+  }
+
+  public captureScreen(): void {
+    const data = document.querySelector('#content') as HTMLElement;
+    const calculatedElements = document.querySelectorAll('mat-list-item') as NodeList;
+
+    html2canvas(data).then((canvas: HTMLCanvasElement) => {
+      const contentDataURL = canvas.toDataURL('image/png');
+
+      const pdf = new jsPDF('p', 'pt', 'a4');
+
+      const imgWidth = 560;
+      const imgHeight = (canvas.height * imgWidth) / canvas.width;
+
+      // pdf.text("Mortgage Expense Calculations", 20, 20);
+      pdf.addImage(contentDataURL, 'PNG', 20, 40, imgWidth, imgHeight, '', 'FAST');
+      pdf.save('Mortgage Expense Calculation.pdf');
+    });
   }
 }
