+++ src/app/components/calculate/results/results.component.html
@@ -207,5 +207,3 @@
     </mat-list-item>
   </mat-list>
 </mat-card>
-
-<!--<button (click)="captureScreen()">Get PDF</button>-->
