+++ src/app/components/calculate/results/results.component.ts
@@ -3,8 +3,6 @@ import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/cor
 import { ExpenseItem, ExpenseVariations, MinMaxModel } from '../../../shared/expense-data.model';
 import { CustomIconService } from '../../../shared/custom-icon.service';
 import { ScrollToConfigOptions, ScrollToService } from '@nicky-lenaers/ngx-scroll-to';
-import * as jsPDF from 'jspdf';
-import html2canvas from 'html2canvas';
 
 @Component({
   selector: 'app-results',
@@ -53,24 +51,6 @@ export class ResultsComponent implements OnInit, OnChanges {
         }
       }
     }
-  }
-
-  public captureScreen(): void {
-    const data = document.querySelector('#content') as HTMLElement;
-    const calculatedElements = document.querySelectorAll('mat-list-item') as NodeList;
-
-    html2canvas(data).then((canvas: HTMLCanvasElement) => {
-      const contentDataURL = canvas.toDataURL('image/png');
-
-      const pdf = new jsPDF('p', 'pt', 'a4');
-
-      const imgWidth = 560;
-      const imgHeight = (canvas.height * imgWidth) / canvas.width;
-
-      // pdf.text("Mortgage Expense Calculations", 20, 20);
-      pdf.addImage(contentDataURL, 'PNG', 20, 40, imgWidth, imgHeight, '', 'FAST');
-      pdf.save('Mortgage Expense Calculation.pdf');
-    });
   }
 
   triggerScrollToEnd(): void {
